backend/requirements.txt
crewai==0.22.5
crewai[tools]
langchain_openai
langchain_community
python-dotenv
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0